// src/app/models/ParichayUser.ts
export interface ParichayUser {
    id:string;
    userName: string;
    browserId: string;
    parichayId: string;
    sessionId: string;
    localTokenId: string;
    firstName: string;
    lastName: string;
    mobileNo: string;
    city: string;
    employeeCode: string;
    fullName: string;
    designation: string;
    email: string;
  }
  